safari-534.53-branch/LayoutTests/http/tests/security/xssAuditor/base-href-direct-expected.txt

ALERT: /XSS/
We allow direct injections into base tags to reduce false positives.
